Homework 5 for CSE240
Creates a list of students for a school. A user is given the option to
add a new student to the list, search for the standard of a student on
the list, or print the list of students. When adding a new student to
the list, the user is prompted for the student’s name, gender, standard,
roll number, and tuition fee. Can save list to a file and it will load
students at the beginning of the program.